Add files via upload
I created this script at my current company to check the margins of a specified brand before a brand campaign. It helps in determining what discount level to set for the campaign. The script loads data via SQL queries from the database and a Google Sheets file, then it creates some views and pivots and loads them into a Google Sheets file for further analysis.